Create hunts pages that render conditionally based on number of hunts
