+++ frontend/src/components/app.jsx
@@ -12,7 +12,7 @@ import Splash from './main/splash';
 import SessionFormContainer from './session/session_form_container';
 import DashboardContainer from './dashboard/dashboard_contianer';
 import HuntsIndexContainer from './hunt/hunts_index_container';
-import ChallengeIndexContainer from './challenge/challenge_index_container';
+import ChallengeIndexPa from './challenge/challenge_index_container';
 
 const App = () => (
     <div className='app'>
@@ -22,7 +22,7 @@ const App = () => (
                 <AuthRoute exact path={AUTH} component={SessionFormContainer} />
                 <ProtectedRoute exact path={DASH} component={DashboardContainer} />
                 <ProtectedRoute exact path={CREATE} component={CreateFormContainer} />
-                <ProtectedRoute exact path={INDEX} component={HuntsIndexContainer} />
+                <ProtectedRoute exact path={INDEX} component={HuntsIndexPageContainer} />
                 <ProtectedRoute exact path={CINDEX} component={ChallengeIndexContainer} />
             </Switch>
          <Footer />
