+++ frontend/src/components/hunt/hunt_collection_item.jsx
@@ -7,16 +7,27 @@ class HuntCollectionItem extends React.Component {
 
   render() {
       return(
-      <div className="mini-cards">
-        <div className="mini-card mini-one">
-
-        </div>
-        <div className="mini-card mini-two">
-        Collection
-        </div>
-        <div className="mini-card mini-three">
-        </div>
-      </div>
+        {this.props.hunts ? 
+          <div className="mini-cards">
+            <div className="mini-card mini-one">
+            </div>
+            <div className="mini-card mini-two">
+            Collection
+            </div>
+            <div className="mini-card mini-three">
+            </div>
+          </div> 
+          : 
+          <div className="mini-cards">
+            <div className="mini-card mini-one">
+            </div>
+            <div className="mini-card mini-two">
+            Collection
+            </div>
+            <div className="mini-card mini-three">
+            </div>
+          </div> 
+        }
       )
   }
 }
