+++ frontend/src/stylesheets/components/challenge/_challenge_index_inactive.scss
