+++ frontend/src/stylesheets/components/hunt/_create_form.scss
@@ -78,7 +78,6 @@
   overflow-x: hidden;
 }
 
-
 .create-form-img-preview {
   height: 200px;
   width: 200px;
@@ -89,9 +88,6 @@
 .hunt-photo {
   display: flex;
   justify-content: center;
-}
-
-.create-from-img-delete{
 }
 
 .create-form-body-right {
